x Removing tabulations x Correcting compilation issues with some options x Adding some compilation options
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@3853 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Audio/File_Aac_GeneralAudio_Sbr.cpp
@@ -450,7 +450,7 @@ void File_Aac::sbr_single_channel_element()
     Get_SB (bs_extended_data,                                   "bs_extended_data[0]");
     if (bs_extended_data) {
         Get_S1 (4,bs_extension_size,                            "bs_extension_size");
-        int cnt=bs_extension_size;
+        size_t cnt=bs_extension_size;
         if (cnt==15)
         {
             Get_S1 (8, bs_esc_count,                            "bs_esc_count");
@@ -600,7 +600,7 @@ void File_Aac::sbr_channel_pair_element()
     Get_SB (bs_extended_data,                                   "bs_extended_data");
     if (bs_extended_data) {
         Get_S1(4,bs_extension_size,                             "bs_extension_size");
-        int cnt = bs_extension_size;
+        size_t cnt = bs_extension_size;
         if (cnt == 15)
         {
             Get_S1(8,bs_esc_count,                              "bs_esc_count");
